Emotion-Detection-1/requirements.txt
anyio
argon2-cffi
arrow
async-lru
attrs
bcrypt
beautifulsoup4
bleach
certifi
cffi
chardet
charset-normalizer
click
colorama
contourpy
cryptography
cycler
decorator
defusedxml
distro
executing
fastjsonschema
filelock
fonttools
fsspec
h11
httpcore
httpx
idna
ipykernel
ipython
ipywidgets
jedi
jinja2
joblib
json5
jsonschema
jupyter
jupyterlab
kaggle
kagglehub
kiwisolver
lark
markdown-it-py
markupsafe
matplotlib
mistune
mpmath
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
numpy
opencv-python
packaging
pillow
prometheus_client
protobuf
psutil
pygments
PyJWT
pyOpenSSL
pyparsing
python-dateutil
python-slugify
pytz
PyYAML
pyzmq
requests
rich
scikit-learn
scipy
sympy
terminado
threadpoolctl
tinycss2
torch
torchvision
tornado
tqdm
traitlets
twisted
typing_extensions
urllib3
webcolors
websocket-client
widgetsnbextension
efficientnet_pytorch
transforms
